+++ client/state/media/thunks/index.js
@@ -1,0 +1,2 @@
+export { addMedia } from './add-media';
+export { uploadSiteIcon } from './upload-site-icon';
